Simpler execution of all SQL scripts
Batch and shell scripts added to run all of the SQL scripts in sequence.

Indices are now created once the tables are populated, rather than during table creation. This greatly improves the runtime of the SQL in batch mode. It was either this solution or running an ANALYZE TABLE after the INSERT.